+++ src/layouts/sample/ImageArea/ImageArea.html
