+++ src/layouts/sample/ImageArea/ImageArea.scss
@@ -1,4 +1,4 @@
-@import '../../palette';
+@import '../../../palette';
 
 .image-area {
   .sample-list {
