+++ src/layouts/sample/LazyArea/lazy-content.js
